refactor: 責務分割とディレクトリ構造の整理を実施\n\n- データ取得、処理、プロット管理のモジュールを分割\n- 型定義と定数をutilsディレクトリに集約\n- テストを追加して品質を向上
