--- docs/development.md
@@ -1,90 +1,0 @@
-# 開発ガイドライン
-
-## 開発環境
-
-### 必要なツール
-- Python 3.8以上
-- Visual Studio Build Tools（C++コンパイラ）
-- Git
-
-### 推奨IDE
-- Visual Studio Code
-  - Python拡張機能
-  - Git拡張機能
-
-## コーディング規約
-
-### Pythonコード規約
-- PEP 8に準拠
-- インデント: スペース4つ
-- 行の最大長: 88文字
-- 文字エンコーディング: UTF-8
-
-### コメント規約
-- 関数やクラスには必ずドキュメンテーション文字列を記述
-- 複雑なロジックには適切なインラインコメントを追加
-- コメントは日本語で記述
-
-### 命名規則
-- クラス名: UpperCamelCase
-- 関数名・変数名: snake_case
-- 定数: UPPER_SNAKE_CASE
-- プライベート変数・関数: _snake_case（アンダースコアプレフィックス）
-
-## ブランチ戦略
-
-### ブランチ命名規則
-- 機能開発: `feature/機能名`
-- バグ修正: `fix/バグ内容`
-- リファクタリング: `refactor/内容`
-- ドキュメント: `docs/内容`
-
-### 開発フロー
-1. 新規ブランチを作成
-2. 変更を実装
-3. テストを実行
-4. プルリクエストを作成
-5. レビュー後にマージ
-
-## テスト
-
-### テスト方針
-- ユニットテスト: 各関数・クラスの動作確認
-- 統合テスト: コンポーネント間の連携確認
-- E2Eテスト: ユーザーシナリオベースのテスト
-
-### テストツール
-- unittest: Pythonの標準テストフレームワーク
-- pytest: より高度なテスト機能を提供
-
-## デプロイ
-
-### デプロイ環境
-- ローカル開発環境
-- テスト環境（必要に応じて）
-- 本番環境
-
-### デプロイ手順
-1. 依存パッケージのインストール確認
-2. テストの実行
-3. アプリケーションの起動
-4. 動作確認
-
-## セキュリティ
-
-### セキュリティガイドライン
-- APIキーなどの機密情報は環境変数で管理
-- 依存パッケージの定期的な更新
-- ユーザー入力のバリデーション実施
-
-## パフォーマンス
-
-### パフォーマンス目標
-- ページ読み込み: 3秒以内
-- データ取得: 5秒以内
-- メモリ使用量の最適化
-
-### 最適化戦略
-- データのキャッシュ活用
-- 非同期処理の活用
-- 効率的なデータ構造の使用
